Make the `LoggingConnection` impl fully generic
The old impl was backend specific, so replace that with a generic impl
that just supports every connection type that somehow implements
`diesel::connection::Connection`
Also migrate the crate to the 2018 edition
